CIP20150217 Dynamic Property Lookup
It has been requested by users to be able to lookup the key of an entity
or a map using a dynamically computed String value as the key.
Currently Cypher does not provide this.  This CIP suggests the
introduction of new syntax (`n["key"]``) to enable this.  Additionally,
this CIP proposes to add a new function (`keys()`) for iterating
the keys of any map or entity.