Filter out legacy rules during resolve step
The openCypher grammar should reflect which parts of the Cypher language
that are standardized. To allow tool authors to still write tools for the de facto
grammar used by the only known Cypher implementation (Neo4j), rules are
not removed, but marked.

The default use is to ignore legacy rules, but supplying an option
-DINCLUDE_LEGACY="true" (or set the java property directly) will include
legacy rules.