Align side effects with standard definition
As of PR #236, side effects are properly defined as the difference in
returned results between a set of Cypher queries, executed before and
after the query in a TCK scenario. This commit aligns the specified side
effects in many scenarios to these new definitions.

Worth noting at this point is that the `labels` metric measures the
number of distinct labels present in the graph, and not the number of
nodes for all labels, as was the initial design.