Equality not defined from comparability. Changes around NaN.
This changes the definitions for equality and comparability such that
we have two basic definitions `=` and `<` and all other operators are
defined using those. This has the advantage that equality between
different types is now `false` instead of `null`.

This also changes equality and comparability around NaN to be compatible
with IEEE 754. The changes include that comparing (<, <=, >, >=) numbers
to NaN is always false, equality between a number and NaN is always
false and `<>` between a number and Nan is always true.