Makes directory-encoded feature categories available in the TCK API.
The scenario case class gains a field `category`` which is a list of strings that contains the top-down category hierarchy of the scenario's feature file. Example: Any scenario from `clauses/match/Match1.feature` will have a list `List("clauses", "match")`.

This allows to filter scenarios based on their category.

The change also comprises a test for the new functionality.