Restructure Scala Dependencies
For the 2.13 cross build to work the Scala dependencies cannot be in the top level .pom as this will not work in a dependency hierarchy unless everything has a Scala version. To fix, move all the Scala dependencies under the scala modules so that the non-Scala modules can stay as they are. tck-reporting has direct Scala dependencies so no choice there.