Update expected errors in TCK
Background:
When referring to a concrete variable in an aggregation expresion in ORDER BY,
e.g. referring to `you` in

```
WITH count(you.age) AS agg
ORDER BY me.age + count(you.age)
```

there are 2 errors:
* `you` is not in scope ("UndefinedVariable")
* It is not OK to use an aggregation function in ORDER BY ("AmbiguousAggregationExpression")

We are changing the way of our SemanticAnalysis such that the first error takes precedence.